requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Markdown==3.5.2
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
